Break RedshiftIntegrationSuite into smaller suites; refactor to remove some redundancy
This patch breaks the monolithic `RedshiftIntegrationSuite` into multiple smaller suites, paving the way to run subsets of these tests with different configurations.

Author: Josh Rosen <joshrosen@databricks.com>

Closes #286 from JoshRosen/remove-dead-code-in-test-suite.
