+++ src/it/scala/com/databricks/spark/redshift/IntegrationSuiteBase.scala
@@ -25,7 +25,7 @@ import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.{Path, FileSystem}
 import org.apache.hadoop.fs.s3native.NativeS3FileSystem
 import org.apache.spark.SparkContext
-import org.apache.spark.sql.{DataFrame, SaveMode, SQLContext}
+import org.apache.spark.sql._
 import org.apache.spark.sql.hive.test.TestHiveContext
 import org.apache.spark.sql.types.StructType
 import org.scalatest.{BeforeAndAfterEach, BeforeAndAfterAll, Matchers}
@@ -120,6 +120,67 @@ trait IntegrationSuiteBase
   override protected def beforeEach(): Unit = {
     super.beforeEach()
     sqlContext = new TestHiveContext(sc, loadTestTables = false)
+  }
+
+  /**
+   * Create a new DataFrameReader using common options for reading from Redshift.
+   */
+  protected def read: DataFrameReader = {
+    sqlContext.read
+      .format("com.databricks.spark.redshift")
+      .option("url", jdbcUrl)
+      .option("tempdir", tempDir)
+  }
+  /**
+   * Create a new DataFrameWriter using common options for writing to Redshift.
+   */
+  protected def write(df: DataFrame): DataFrameWriter[Row] = {
+    df.write
+      .format("com.databricks.spark.redshift")
+      .option("url", jdbcUrl)
+      .option("tempdir", tempDir)
+  }
+
+  protected def createTestDataInRedshift(tableName: String): Unit = {
+    conn.createStatement().executeUpdate(
+      s"""
+         |create table $tableName (
+         |testbyte int2,
+         |testbool boolean,
+         |testdate date,
+         |testdouble float8,
+         |testfloat float4,
+         |testint int4,
+         |testlong int8,
+         |testshort int2,
+         |teststring varchar(256),
+         |testtimestamp timestamp
+         |)
+      """.stripMargin
+    )
+    // scalastyle:off
+    conn.createStatement().executeUpdate(
+      s"""
+         |insert into $tableName values
+         |(null, null, null, null, null, null, null, null, null, null),
+         |(0, null, '2015-07-03', 0.0, -1.0, 4141214, 1239012341823719, null, 'f', '2015-07-03 00:00:00.000'),
+         |(0, false, null, -1234152.12312498, 100000.0, null, 1239012341823719, 24, '___|_123', null),
+         |(1, false, '2015-07-02', 0.0, 0.0, 42, 1239012341823719, -13, 'asdf', '2015-07-02 00:00:00.000'),
+         |(1, true, '2015-07-01', 1234152.12312498, 1.0, 42, 1239012341823719, 23, 'Unicode''s樂趣', '2015-07-01 00:00:00.001')
+         """.stripMargin
+    )
+    // scalastyle:on
+    conn.commit()
+  }
+
+  protected def withTempRedshiftTable[T](namePrefix: String)(body: String => T): T = {
+    val tableName = s"$namePrefix$randomSuffix"
+    try {
+      body(tableName)
+    } finally {
+      conn.prepareStatement(s"drop table if exists $tableName").executeUpdate()
+      conn.commit()
+    }
   }
 
   /**
